Use RL.jl v0.11

+++ src/DDDC2023/env_helpers.jl
@@ -5,7 +5,7 @@ Construct a lookup table from action space to the state space.
 """
 function construct_DDDC_state_space_lookup(action_space, n_prices)
     @assert length(action_space) == n_prices^2 * 4
-    state_space_lookup = reshape(Int16.(1:length(action_space)), n_prices, n_prices, 2, 2)
+    state_space_lookup = reshape(Int64.(1:length(action_space)), n_prices, n_prices, 2, 2)
     return state_space_lookup
 end
